HOTT-3032: Rename suggestion service
I'm doing this because their is only one service and its not specific to
any api
